Just a little tweak
while easy to handwave, the line "it might just end up sending fragments of the ruin across the stars!" doesn't really make much sense from both a logical and lore perspective, I won't nerd on the lore side of it, but sublight bits of organic matter really isn't much to be concerned of. If anything it'll be easy to track down and eliminate. What's more at stake is that something that's already been obliterated will be reduced even more (kinda like kicking a person when they're down), and the fact that someone as insane as the Survivor probably won't stop with """cleansing""" earth, makes for far more valid and logical concerns.